REF start a major refactor to enable resolvo+rattler
