+++ conda_forge_feedstock_check_solvable/__init__.py
@@ -1,3 +1,4 @@
 # flake8: noqa
 from ._version import __version__
-from .mamba_solver import MambaSolver, is_recipe_solvable
+from .check_solvable import is_recipe_solvable
+from .mamba_solver import MambaSolver
